refactor: move applyImageTransformations to utils as imageTransformations.ts
