+++ src/shared/utils/index.ts
@@ -4,6 +4,7 @@ export * from './formatDate'
 export * from './getBaseUrl'
 export * from './handleErrorResponse'
 export * from './hooks'
+export * from './imageTransformations'
 export * from './imageUtils'
 export * from './isValidErrorResponse'
 export * from './storage'
